BugFix: Terminal render too many times causing performance freeze
